Move NLTE data back
